checkpoint: extracts closure
