+++ src/library/ShimmedStabilityAIClient/conversions/exports.ts
@@ -1,1 +1,3 @@
 export * from './toShortfinRequestBodyPrompt';
+
+export * from './toShortfinRequestBody';
