+++ src/library/ShimmedStabilityAIClient/conversions/toShortfinRequestBody.ts
@@ -1,0 +1,41 @@
+import type {
+  GenerateFromTextRequest,
+} from 'stabilityai-client-typescript/models/operations';
+
+import type Shortfin from '@/library/Shortfin';
+
+import {
+  toShortfinRequestBodyPrompt,
+} from './toShortfinRequestBodyPrompt';
+
+const toShortfinRequestBody = (
+  givenRequestBody: GenerateFromTextRequest['textToImageRequestBody'],
+): Shortfin.TextToImage.SDXL.Client.Request.Body | null => {
+  if (
+    (givenRequestBody.height === undefined)
+    || (givenRequestBody.width === undefined)
+    || (givenRequestBody.steps === undefined)
+    || (givenRequestBody.cfgScale === undefined)
+    || (givenRequestBody.seed === undefined)
+  ) return null;
+
+  const derivedShortfinRequestBody = {
+    prompt: toShortfinRequestBodyPrompt(givenRequestBody.textPrompts, {
+      weight: 1,
+    }),
+    neg_prompt: toShortfinRequestBodyPrompt(givenRequestBody.textPrompts, {
+      weight: -1,
+    }),
+    height        : givenRequestBody.height,
+    width         : givenRequestBody.width,
+    steps         : givenRequestBody.steps,
+    guidance_scale: givenRequestBody.cfgScale,
+    seed          : givenRequestBody.seed,
+  };
+
+  return derivedShortfinRequestBody;
+};
+
+export {
+  toShortfinRequestBody,
+};
